move delete_key_recursive to coding_utils.py
